Standardize verb usage within the module (*Breaking change*)
This attempts to rectify some improper verb usage in the module
based on the explanations of intended verb usage from [previous
PowerShell documentation](https://web.archive.org/web/20171222220053/https://msdn.microsoft.com/en-us/library/windows/desktop/ms714428(v=vs.85).aspx).

* We're standardizing on the following pattern for most object actions:
    `Get` / `Set` / `New` / `Remove`

* We will continue to alias `Remove-*` as `Delete-*`.

* When possible, this change attempts to avoid breaking changes by
  re-aliasing the newly named functions with their previous names.
  This was not possible in one instance, hence this is still a breaking
  change (although based on telemetry, it should be minimally impacting).

Result:
 * `Update-GitHubCurrentUser` -> `Set-GitHubProfile` `[Alias('Update-GitHubCurrentUser')]`
 * `Update-GitHubIssue` -> `Set-GitHubIssue`  `[Alias('Update-GitHubIssue')]`
 * `Update-GitHubRepository` -> `Set-GitHubRepository`  `[Alias('Update-GitHubRepository')]`
 * [breaking] `Update-GitHubLabel` -> `Set-GitHubLabel`  `[Alias('Update-GitHubLabel')]`
 * [breaking] `Set-GitHubLabel` -> `Restore-GitHubLabel` `<no alias due to above>`

Changing an existing label has much more regular usage than replacing
all of the labels in a repository, hence allowing the _new_
`Set-GitHubLabel` to keep the alias of `Update-GitHubLabel`.

Our usage of the `Set-*` verb in general is a bit arguable based on
the documentation ... in theory `Edit-*` might be a better fit since
we're _editing_ aspects of an object as opposed to _replacing_ the entire
content of an object.  However, I think `Set-*` _feels_ ok in this module.
We're _setting_ the state of these objects.  Again...arguable, but this
is a much smaller breaking change to get to a consistent terminology state.