Improve performance of Get-GitHubContent
The call to `ConvertTo-SmarterObject` was adding a significant amount
of time to the execution of Get-GitHubContent.

```powershell
    Set-GitHubConfiguration -DisableSmarterObjects:$false
    Measure-Command { $c = Get-GitHubContent -OwnerName microsoft -RepositoryName PowerShellForGitHub -Path README.md }
    # This averages to be around 1.2 seconds

    Set-GitHubConfiguration -DisableSmarterObjects:$true
    Measure-Command { $c = Get-GitHubContent -OwnerName microsoft -RepositoryName PowerShellForGitHub -Path README.md }
    # This averages to be around 14 seconds
```

The reasoning behind this was that the return of `Invoke-WebRequest`
for this repo's `README.md` was an `[Object[]]` that contained 22,000+
entries (one byte per entry).  Piping in 22,000 objects and just writing
each of them out to via `Write-Output` took a _lot_ of time.

The fix here was to skip trying to convert the result to a smarter object
if it wasn't going to be a convertible object in the first place.

This doesn't end up saving much time with the execution of
`Tests/GitHubContents.tests.ps1` because the file being returned
is so small, but this should have some nice user experience improvements
in practical usage.