Fix usage of Resolve-RepositoryElements -DisableValidation
There are only two legitimate instances in the module as of today
where we should be disabling the validation that Resolve-RepositoryElements
provides.  All other instances were likely introduced due to mistaken
copy/paste issues.

The only reason we'd want to disable the validation is if there are
multiple parametersets that need to be considered for the function,
where at least one _would_ require `OwnerName`/`RepositoryName` while
at least one other one _would not_.

This would have meant that we would have attempted to call those
GitHub endpoints with incomplete path information which would have
most likely resulted in an error that we could have caught sooner
without needing to invoke a web request.

I've fixed up all of these incorrect instances, and added a comment
to the two remaining ones to remind others later on why that switch
is being used, and help prevent incorrect usage in the future.