Improve CI speed by changing repo targeted by Forks tests
Pull #251 changed the repo targeted by the forks tests from
`microsoft/PowerShellForGitHub` to `octocat/Hello-World` in order
to prevent the accidental deletion of real forks when running the UT's
locally against your own account.

In practice, this almost doubled the execution time of the UT's,
because the execution time of the Forks API's took so much longer
against a repo with 1400+ forks.

Changing the test over to use a repo that currently only has 39 forks
(several orders of magnitude fewer).