Fix handling of numerical configuration properties across PS5 and PS7
ConvertFrom-Json defaults numbers to different types depending on
the PowerShell version.

PS5 defaults numbers to [Int32] while PS7 defaults numbers to [int64],
but both versions default to [Int32] in [PSCustomObjects].

Due to this, if you had a number config value saved, on PS7 it would be
ignored and the default value would be used because the expected and
actual types wouldn't match.

I've updated `Resolve-PropertyValue` to have equivalence logic for
number types to avoid this issue.