Attempting to increase reliability of some Pester tests - take 2
This is re-implementing #265 in a more robust way, thanks to a
suggestion from @X-Guardian.

Instead of adding in one-off sleeps throughout the test codebase,
there is now a new configuration value that will allow us to insert
a delay before returning the result of _any_ state change request.

This should ideally consistently add reliability throughout the entire
test codebase.