Removing NoStatus from the module
The `NoStatus` parameter (and `DefaultNoStatus` config value) were
deprecated as part of #253 when we stopped showing status except for
multi-page requests that exceeded some minimum number of pages.

At that time, `NoStatus` (and `DefaultNoStatus`) were not removed in
order to minimize the churn to the module.  However, keeping it in
is adding unnecessary complexity to the module as we continue to
expand what the module can do.

This change removes `NoStatu` (and `DefaultNoStatus`) from the module
entirely.  The only impact that this may cause is with users who are
currently using one (or both) of them.  This breaking change impact
should be minimal, but its best for the breaking change to be part of
the coming release which already has a number of other breaking changes
as well, as opposed to having more breaking changes in a successive
release.