Fix default log location when documents folder is unavailable
In Azure Function Apps, the special `MyDocuments` folder resolves to
an empty string, which means that the default `logPath` gets stored as
an empty string.  This causes problems when `Write-Log` attempts to
write anything out.

In this scenario, we'll fall back to using the temp folder and writing
a warning out to the user letting them know this has happened.

Resolves #282