Fix gist test validating the since parameter
octocat recently updated all of its gists, starting on 2/24/2021.
The test previously would query for gists that were updated since 2016,
but as of 2/24/2021, that included all of them.

The test has been made more robust by choosing the most recent date of
all the gists and then requerying, which should hopefully now always
guarantee that we'll get a reduced set of gists, unless all of the gists
are suddenly updated in the future at the exact same second.