Fix coalesced Team permission property
When retrieving the permissions granted to a team for a repository,
the API returns an array of booleans indicating what individual
permissions the team currently has.

| name | permissions granted |
| ---- | ------------------ |
| pull | pull |
| triage | pull, triage |
| push | pull, triage, push |
| maintain | pull, triage, push, maintain |
| admin | pull, triage, push, maintain, admin |

As you can see, each permissions builds on top of the previous
permission.  When adding extra properties to this returned object
via `Add-GitHubRepositoryTeamPermissionAdditionalProperties`, we
also add a new `permission` property that attempts to represent the
collection of permissions with their current top-permission level.

The logic however incorrectly favored `push` over `maintain` and thus
would always said that a Team had `push` permission when in fact it
may have actually had `maintain` permission.

This simple change corrects that ordering mistake.